+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+* Update examples and documentation to use cortex-m-rtfm v0.5
+
+
 ## rmicrobit 0.4.0 (2019-11-03)
 
 * First release to crates.io